AccessBackEnd/requirements.txt
Flask>=3.0,<4.0
Flask-Cors>=4.0,<5.0
Flask-JWT-Extended>=4.6,<5.0
Flask-Login>=0.6,<1.0
Flask-Migrate>=4.0,<5.0
Flask-SQLAlchemy>=3.1,<4.0
pytest>=8.0,<9.0
huggingface-hub>=0.24,<1.0
transformers>=4.44,<5.0
langchain-core>=0.2,<1.0
langchain-community>=0.2,<1.0
torch>=2.2,<3.0
marshmallow>=3.22,<4.0